requirements.txt
Package                  Version
------------------------ ----------
blinker                  1.9.0
certifi                  2025.1.31
cffi                     1.17.1
charset-normalizer       3.4.1
click                    8.1.8
filelock                 3.17.0
Flask                    3.1.0
flask-cors               5.0.1
fsspec                   2025.3.0
huggingface-hub          0.29.3
idna                     3.10
itsdangerous             2.2.0
Jinja2                   3.1.6
llvmlite                 0.44.0
MarkupSafe               3.0.2
more-itertools           10.6.0
mpmath                   1.3.0
networkx                 3.4.2
numba                    0.61.0
numpy                    2.1.3
nvidia-cublas-cu12       12.4.5.8
nvidia-cuda-cupti-cu12   12.4.127
nvidia-cuda-nvrtc-cu12   12.4.127
nvidia-cuda-runtime-cu12 12.4.127
nvidia-cudnn-cu12        9.1.0.70
nvidia-cufft-cu12        11.2.1.3
nvidia-curand-cu12       10.3.5.147
nvidia-cusolver-cu12     11.6.1.9
nvidia-cusparse-cu12     12.3.1.170
nvidia-cusparselt-cu12   0.6.2
nvidia-nccl-cu12         2.21.5
nvidia-nvjitlink-cu12    12.4.127
nvidia-nvtx-cu12         12.4.127
openai-whisper           20240930
packaging                24.2
pillow                   11.1.0
pip                      24.0
pycparser                2.22
pytesseract              0.3.13
PyYAML                   6.0.2
regex                    2024.11.6
requests                 2.32.3
setuptools               76.0.0
six                      1.17.0
soundfile                0.13.1
sympy                    1.13.1
tenacity                 9.0.0
tiktoken                 0.9.0
torch                    2.6.0
tqdm                     4.67.1
triton                   3.2.0
typing_extensions        4.12.2
urllib3                  2.3.0
Werkzeug                 3.1.3
